// src/enums/error.enum.ts
export enum ErrorEnum {
  GENERIC_ERROR = 'GENERIC_ERROR',
  NOT_FOUND_ERROR = 'NOT_FOUND_ERROR',
  SERVICE_ERROR = 'SERVICE_ERROR',
  NOT_AUTHENTICATED_ERROR = 'NOT_AUTHENTICATED_ERROR',
  NOT_AUTHORIZED_ERROR = 'NOT_AUTHORIZED_ERROR',
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  CLIENT_ERROR = 'CLIENT_ERROR',
  CANCELED_ERROR = 'CANCELED_ERROR',
}